[PATCH] misc fixes
- add init_timer in bttv driver

- remove duplicated init_timer() in ncpfs.

- remove noisy printk's from 3c59x.c

- sparc64 compile fix with CONFIG_HUGETLBPAGE=y - htlbpage_max is now an
  int

+++ arch/sparc64/mm/init.c
@@ -1696,7 +1696,7 @@ static void __init taint_real_pages(void)
 
 #ifdef CONFIG_HUGETLB_PAGE
 long htlbpagemem = 0;
-long htlbpage_max;
+int htlbpage_max;
 long htlbzone_pages;
 extern struct list_head htlbpage_freelist;
 #endif
